+++ bootpack.h
@@ -52,6 +52,7 @@ struct BOOTINFO
 };
 
 #define ADR_BOOTINFO 0x00000ff0
+
 /* dsctbl.c */
 struct SEGMENT_DESCRIPTOR
 {
@@ -67,9 +68,9 @@ struct GATE_DESCRIPTOR
     short offset_high;
 };
 
+void init_gdtidt(void);
 void set_segmdesc(struct SEGMENT_DESCRIPTOR *sd, unsigned int limit, int base, int ar);
 void set_gatedesc(struct GATE_DESCRIPTOR *gd, int offset, int selector, int ar);
-void init_gdtidt(void);
 
 #define ADR_IDT 0x0026f800
 #define LIMIT_IDT 0x000007ff
@@ -81,11 +82,9 @@ void init_gdtidt(void);
 #define AR_CODE32_ER 0x409a
 #define AR_INTGATE32 0x0008e
 
-/* init.c */
+/* int.c */
 
 void init_pic(void);
-void inthandler21(int *esp);
-void inthandler2c(int *esp);
 #define PIC0_ICW1 0x0020
 #define PIC0_OCW2 0x0020
 #define PIC0_IMR 0x0021
@@ -109,3 +108,22 @@ void fifo8_init(struct FIFO8 *fifo, int size, unsigned char *buf);
 int fifo8_put(struct FIFO8 *fifo, unsigned char data);
 int fifo8_get(struct FIFO8 *fifo);
 int fifo8_status(struct FIFO8 *fifo);
+
+/* keyboard.c */
+void inthandler21(int *esp);
+void wait_KBC_sendready(void);
+void init_keyboard(void);
+extern struct FIFO8 keyfifo;
+#define PORT_KEYDAT 0x0060
+#define PORT_KEYCMD 0x0060
+
+/* mouse.c */
+struct MOUSE_DEC
+{
+    unsigned char buf[3], phase;
+    int x, y, btn;
+};
+void inthandler2c(int *esp);
+void enable_mouse(struct MOUSE_DEC *mdec);
+int mouse_decode(struct MOUSE_DEC *mdec, unsigned char data);
+extern struct FIFO8 mousefifo;
