+++ int.c
@@ -18,28 +18,3 @@ void init_pic(void)
     io_out8(PIC0_IMR, 0xfb);
     io_out8(PIC1_IMR, 0xff);
 }
-
-#define PORT_KEYDAT 0x0060
-
-struct FIFO8 keyfifo;
-/* ps/2 keyboard 割り込み */
-void inthandler21(int *esp)
-{
-    unsigned char data;
-    io_out8(PIC0_OCW2, 0x61); /* IRQ-01受付完了をPIC0に通知 */
-    data = io_in8(PORT_KEYDAT);
-    fifo8_put(&keyfifo, data);
-    return;
-}
-
-struct FIFO8 mousefifo;
-/* ps/2 mouse 割り込み */
-void inthandler2c(int *esp)
-{
-    unsigned char data;
-    io_out8(PIC1_OCW2, 0x64); /* IRQ-12受付完了をPIC1に通知 */
-    io_out8(PIC0_OCW2, 0x62); /* IRQ-02受付完了をPIC0に通知 */
-    data = io_in8(PORT_KEYDAT);
-    fifo8_put(&mousefifo, data);
-    return;
-}
